+++ tests/test_imports.py
@@ -1,3 +1,0 @@
-def test_imports_for_ndarray_issues():
-    # this test does nothing, but it raises warnings within pytest!
-    import pycbc.psd