+++ CHANGELOG.md
@@ -12,7 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     post-Newtonian expressions. 
     (Right now, this is still [not smooth](https://github.com/jacopok/mlgw_bns/issues/36).)
 - Mention of this changelog in the README
-- Reference documentation about the mathematical details of higher order modes.
+- Reference documentation about the mathematical details of higher order modes
+- Removed dependence on `pycbc` for PSD computations (see [this PR](https://github.com/jacopok/mlgw_bns/pull/38)): 
+    this significantly decreases the dependency load of the package
+
+### Removed
+
+- Python 3.7 support
 
 ## [0.10.2] - 2022-07-01